Refactor database class
